+++ src/pages/Dashboard/components/AccountList/AccountList.tsx
@@ -1,10 +1,15 @@
 import { Stack } from '@chakra-ui/layout'
+import { Button } from '@chakra-ui/react'
 import { NetworkTypes } from '@shapeshiftoss/types'
 import { useEffect } from 'react'
 import { useMemo } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { AccountRow } from 'components/AccountRow/AccountRow'
 import { LoadingRow } from 'components/AccountRow/LoadingRow'
+import { Card } from 'components/Card/Card'
+import { RawText } from 'components/Text'
+import { useModal } from 'context/ModalProvider/ModalProvider'
+import { useWallet, WalletActions } from 'context/WalletProvider/WalletProvider'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { usePortfolio } from 'pages/Dashboard/contexts/PortfolioContext'
 import { sortByFiat } from 'pages/Dashboard/helpers/sortByFiat/sortByFiat'
@@ -16,6 +21,11 @@ export const AccountList = ({ loading }: { loading?: boolean }) => {
   const assets = useSelector((state: ReduxState) => state.assets)
   const marketData = useSelector((state: ReduxState) => state.marketData.marketData)
   const { balances, totalBalance } = usePortfolio()
+  const { receive } = useModal()
+  const {
+    state: { isConnected },
+    dispatch: walletDispatch
+  } = useWallet()
   const emptyAccounts = new Array(5).fill(null)
 
   useEffect(() => {
@@ -27,6 +37,23 @@ export const AccountList = ({ loading }: { loading?: boolean }) => {
   }, [])
 
   const accountRows = useMemo(() => {
+    const accounts = Object.keys(balances)
+      .sort(sortByFiat({ balances, assets, marketData }))
+      .filter(key => bnOrZero(balances[key].balance).gt(0))
+    if (accounts.length === 0) {
+      const handleWalletModalOpen = () =>
+        walletDispatch({ type: WalletActions.SET_WALLET_MODAL, payload: true })
+      const handleReceiveClick = () => (isConnected ? receive.open({}) : handleWalletModalOpen())
+      return (
+        <Card textAlign='center' py={6} boxShadow='none'>
+          <Card.Body>
+            <RawText>No assets yet.</RawText>
+            <Button onClick={handleReceiveClick}>Receieve some funds</Button>
+          </Card.Body>
+        </Card>
+      )
+    }
+
     return (
       <>
         {Object.keys(balances)
